Started check_yaml.py, checked off project_plan
Started on check_yaml.py, which goes through the setup.yml file and
makes sure the syntax is correct and should raise the chances of success
to deployment CTFd. Needs like 5 lines of code.

Checked off some checkboxes in project_plan. Minimum project is done,
time for extra tasks ;)

Added a check for docker challenge local deployment and added a call for
the function to start. Added a start of check_yaml.py. Moved tz check
into start as to not start on any copy/paste before the setup.yml
file is setup properly.