Documentation, start.sh bug, and notice
Added a clause in the LICENSE to make sure it sticks to forever. Added a
notice in the README to make sure people understand that CTFdeploy
alters some files in CTFd to make sure it works properly. Made sure
start.sh exits with code 0 and when called from other directories it
behaves properly. Added some brief documentation on how CTFdeploy works.