Timezone issues
Depending on what the systems current timezone is set to, CTFd will
change to that zone on setup. This will also add or subtract from the
times set for start and stop. To make sure the time set in setup.yml is
the one displayed in CTFd, an addition/subtraction has to be made
according to the timezone the system is set to.

The problem was tackled using python and figuring out what timezone the
system is set to and if daylight saving has to be accounted for. Also,
daylight has to be subtracted if in a "positive" timezone and
added if in a "negative" timezone.

The setup has been tested with some different timezones and seem to
deploy as wanted.