Updated README and start.sh
Updated start.sh to handle resetting CTFd better so to not detach
HEAD. It will still reset the files in CTFd but will preserve the use of
committing to CTFd and using it like a normal repository.

Updated README with a warning about committing to CTFd if changes were
made and add '#' instead of sudo for a more universal root sign. Removed
--init and --recursive as these seem unnecessary.